node: removed Parallel() option for test, and merged node_cb_tests with original node_test protocol: added todo for test
